requirements.txt
torch
streamlit
python-dotenv
langchain-core
langchain-community
spacy
spacy-layout
langchain-text-splitters
pydantic
sentence-transformers
neo4j
langchain-together
langgraph
pydot
graphviz
mcp
langchain-mcp-adapters

